Compatibility with OpenShift
OpenShift requires special permissions for setting owner references. We
were unable to identify what specific permission is needed to allow
setting "BlockOwnerDeletion", and we do not require blocking the
deletion of User kind on the deletion of the generated crdedentials
Secret. We decided to simply set to false "BlockOwnerDeletion" to
workaround the problem.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>