Fix endless reconcile issue
Fixes #250

Before this commit, the LastTransitionTime was updated in every
reconcile. This is wrong because LastTransitionTime should only be
updated if the status transitions from one to the other.

This led to endless reconciliations because updating the status leads to
a new resourceVersion resulting in a new reoncile.

With sufficient objects (>100), it takes more than one second between
reconciling the same object. Since the granularity of lastTransitionTime
is 1 second, this ended up in endless updates of lastTranistionTime and
therefore resourceVersion.