Downgrade code-generator to 0.21.x
This was introduced part of
https://github.com/rabbitmq/messaging-topology-operator/pull/87, and in
the latest eventing-rabbitmq changes it became obvious that the versions
started to drift.

This discussion explains the problem best:
https://github.com/rabbitmq/messaging-topology-operator/discussions/256

This holds the rest of the context:
https://github.com/knative-sandbox/eventing-rabbitmq/pull/492#issuecomment-963276439

This change lines up all versions, and the local unit & integration
tests pass. My next step is to check if CI remains green, and if the
change actually works in the context of eventing-rabbitmq, which is how
this started.

Signed-off-by: Gerhard Lazu <glazu@vmware.com>