Fix to running unit-tests locally
MacOS firewall complains with a pop-up notification when something binds
to 0.0.0.0, by making localhost the bind address, we fix the firewall
pop-up problem.

@MarcialRosales and myself were having errors in the unit test
BeforeSuite, starting the local control plane. Arguably those should be
integration tests, given that they require a kube-api and etcd processes
running. In any case, running the tests in parallel seems to cause the
problem because we spawn many kube-api and etcd instances at the same
time, in parallel.
We likely never noticed this in CI because the containers are likely
running in a single vCPU.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>