Support for custom cluster domains
Kubernetes internal domain name can be customised during the cluster
creation. Some certificates are issued with a 'star domain', so that any
subdomain can be validated. In these cases, the Topology Operator needs
a 'hint' about this internal domain, so that it can connect to the
correct URI to communicate with RabbitMQ.

Related to #233.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>