System tests compatibility with Openshift
Overriding rabbitmq cluster security context
in system-tests when running in Openshift.
RABBITMQ_IMAGE env variable is introduced to override the image
of the rabbitmq cluster created by the system-tests
RABBITMQ_IMAGE_PULL_SECRET is introduced to override
the image pull secret and allow to run the system-tests
on a commercial version.