Update workflow for PRs
Update Kubernetes version to latest available and lowest supported (N-2).
Match Go version with our `go.mod`.

Bump Cert Manager to 1.7.0 because 1.8.0 has breaking changes, and we
have not tested whether we are compatible with those changes.

Sometimes, Cert Manager Pods become ready, however Cert Manager itself
is not ready to serve requests. If Topology Operator is deployed at that
time, Cert Manager will not inject the certificate into the Webhooks,
causing an error of the kind "x509: certificate signed by unknown
authority". The solution is to wait for the API to be ready, using their
new CLI `cmctl`.