Spike: Instrument queue reconciler with OTel
This commit is a working PoC of how to instrument our controllers to
produce open telemetry data. In particular, this commit shows how to
produce traces and spans. This works with Jaeger as a collector.

We could further expand this work to produce Prometheus formatted
metrics, given that we already expose some Prometheus metrics in this
Operator.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>