Run integration tests in parallel
Execution time has cut down from 10 minutes to 1 minute and a few
seconds (!!!)

The solution is to start an API server for each parallel process,
listening on different ports. I tried to start only one API server and
share the API information among other parallel processes, but it was not
possible because the testEnv Environment does not implement any of the
marshall interfaces, so it is not possible to binary encode it. Binary
encoding is the method that Ginkgo uses to share information between
processes in the Synchronized Before/After suites.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>