Pin Go version to 1.20 minor in CI
GO 1.21 creates a non-backwards compatible change in the Go mod file. It
adds a 'toolchain' directive that it's not compatible with <1.21
versions. As a consequence, if a go-1.21 binary modifies the Go mod e.g.
via go mod download, the resulting Go mod won't be compatible with
earlier Go version i.e. 1.20.

The above breaks our CI, because our Go mod gets modified, and the Go
version in the Dockerfile is pinned to 1.20.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>