Remove vuln from unit-tests
There is little value in running vuln in unit-tests. When it complains,
it complains about CVEs in your local Go version, which is
not the same Go binary that will build in CI (we use stable
and oldstable).

The target `git-commit-sha` was a no-op. The logic to determine whether
to add a `-` at the end of the commit SHA was broken, it was always
chosing to add a `-`. Now the `GIT_COMMIT` is always the HEAD commit SHA
+ `-dev` suffix.

Signed-off-by: Aitor Perez Cedres <1515757+Zerpet@users.noreply.github.com>