Fix deletion policy tests
The deletion policy was not being set at all for tests. Why? Because
BeforeEach nodes run BEFORE the JustBeforeEach. Therefore, the
BeforeEach was setting the deletion policy in the spec, and right after
the JustAfterEach in the top container was assigning a "zero" value for
the topology type. This effectively "overrode" the value set for
deletion policy.

This was probably passing because Create/Delete were called one after
the other, and sometimes the deletion behaviour simply skipped because
the object was not found, since it was "too soon" and not returned by
the informer cache.

These tests should have been rightfuly red. The were flaking to green.