Use go run for ginkgo CLI
[Why]
To avoid mismatch version between Ginkgo version in Go mod vs the Ginkgo
CLI installed locally in GOBIN.

The change to the port number in testenv is to avoid using ports outside
of valid range (1-65536) when Ginkgo parallel processes was over 10.
Previously, the process number 10 would render port 81810, which is
outside of the valid range.