Refactor controller tests
Controller tests are flaking way too often. One of the test flakes was
around the "retain" feature. These tests were buggy and should have been
failing consistently; they were flaking to green. The problem was a
create-delete right away, without awaiting for validation on creation.
Sometimes, the create was registered in etcd, and it was causing a
legitimate failure (due to test setup, production code was fine). When
the create did not complete and, in parallel, the deletion executed, it
skipped the deletion (because the object did not exist in etcd yet),
therefore, it skipped the HTTP calls, passing the test assertion
(incorrectly).

The other refactor is to provide better test isolation. Some flakes were
happening, rarely, due to test polution, where the manager for one test
was reconciling the object for another test. After the refactor, each
manager is constrained to its own test suite using label selectors and
namespace isolation. This adds a layer of complexity to writing tests
for controllers, however, the setup can be copy-pasted for new tests.
Hopefully that will alleviate the complexity of writing these tests.