Fix vhost controller test flakes
There was test pollution from tests that set vhost limits. The vhost
limits in RMQ HTTP API are implemented as separate API endpoints. When a
vhost has a limit, the controller declares the vhost first, then
declares the limits. The flake happened when a "limits" tests ran before
the "deletion" tests, and left behind a value for `vhostLimits`
variable. This variable is taken unconditionally to initialise a vhost
variable (as it should). The flake is simply fixed by setting the
`vhostLimits` variable back to `nil` after each "limits" test.

Additionally, the "creation" tests should not delete the objects after
each test suite, because the fake client is not prepared to return
appropriate responses to the delete requests, and that would leave the
managers in an infinite reconcile loop, which can delay manager shutdown
after each test.