Recover owner reference test for User
As part of abc57371b359bf099fc89052a47db7e828258c58, we chopped the test
related to owner references. The system tests verify that objects are
garbage collected, and that proves that an owner reference is set.
However, the system tests do not explicitly verify that the object does
not block the owner deletion. This is required for OpenShift
compatibility.

This commit recovers the owner reference test for User generated secret.