olm: refactor to bundle generation
We are using a "generator" template and applying values on top, instead
of using overlays to modify specific parts of the generator file. The
Makefile is a new approach to script the commands, instead of using
Python in the original OLM dedicated git repo. The future idea is to
fold the OLM manifests repo into this one, for the specific bits of
topology operator.