olm: remove leftover files from refactor
These files are not meant to be checked in. The bundle manifest is
generated on every release. The important files are the templates,
because the OLM build is reproducible. The catalog files are support
files to test the OLM package in CI or locally, and are not meant to be
checked in.